[Style] Convert -webkit-text-size-adjust to use strong style types https://bugs.webkit.org/show_bug.cgi?id=296189
Reviewed by Tim Nguyen.

Split TextSizeAdjust and AutosizeStatus out of TextSizeAdjustment.h and into their
own files and make TextSizeAdjust a strong style type.

* Source/WebCore/CMakeLists.txt:
* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/rendering/TextAutoSizing.cpp:
* Source/WebCore/rendering/style/AutosizeStatus.cpp: Renamed from Source/WebCore/rendering/style/TextSizeAdjustment.cpp.
* Source/WebCore/rendering/style/AutosizeStatus.h: Renamed from Source/WebCore/rendering/style/TextSizeAdjustment.h.
* Source/WebCore/rendering/style/RenderStyle.cpp:
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleInlines.h:
* Source/WebCore/rendering/style/RenderStyleSetters.h:
* Source/WebCore/rendering/style/StyleRareInheritedData.cpp:
* Source/WebCore/rendering/style/StyleRareInheritedData.h:
* Source/WebCore/style/StyleAdjuster.h:
* Source/WebCore/style/StyleBuilderCustom.h:
* Source/WebCore/style/StyleExtractorConverter.h:
* Source/WebCore/style/StyleExtractorSerializer.h:
* Source/WebCore/style/StyleResolveForDocument.cpp:
* Source/WebCore/style/values/size-adjust/StyleTextSizeAdjust.cpp: Added.
* Source/WebCore/style/values/size-adjust/StyleTextSizeAdjust.h: Added.

Canonical link: https://commits.webkit.org/297625@main
